[Refactor] Move animate code to DragAndScale
